Added dynamic updating and settings GUI selection of OBS addresses
From Settings > Devices > Edit Sources there is now the option to select the source from a drop down list populated from the source. The option to enter manually is retained. This will currently only work for OBS and in theory for tricaster, the storage objects used for ATEM and Analog Way are different, a common structure would perhaps help going forward.

OBS Source tally list is updated automatically when the OBS instance adds, renames or deletes. Renames are rippled through, deletes are not.